feat: issuesCard - update time elapsed since issue creation regularly
