+++ hooks/useHydration.ts
@@ -1,0 +1,11 @@
+import { useEffect, useState } from "react";
+
+export function useHydration() {
+  const [hydrated, setHydrated] = useState(false);
+
+  useEffect(() => {
+    setHydrated(true);
+  }, []);
+
+  return hydrated;
+}
